bootstrapping headless browser on actual startup
