+++ runtime/mount/endpoints/agent.py
@@ -20,69 +20,11 @@ from ..entrypoint import (
     start_agent_proc,
     start_headless_browser,
 )
-from ..utils import auth_token_sdk, gql_query, sdk_token
 
 connection_idx = 0
 agent_start_lock = asyncio.Lock()
 
 ConnectionRole = Literal["user", "action_handler", "unknown"]
-
-
-async def _resolve_headless_browser_bootstrap() -> tuple[str, dict[str, str]] | None:
-    if pod_id is None:
-        return None
-
-    resp = await gql_query(
-        auth=auth_token_sdk,
-        query="""
-            query AgentBootstrapPodInfo($podId: BigInt!) {
-                podInfos(filter: { id: { equalTo: $podId } }, first: 1) {
-                    nodes {
-                        plotNotebookId
-                        accountId
-                    }
-                }
-            }
-        """,
-        variables={"podId": pod_id},
-    )
-
-    data = resp.get("data")
-    if not isinstance(data, dict):
-        return None
-
-    pod_infos = data.get("podInfos")
-    if not isinstance(pod_infos, dict):
-        return None
-
-    nodes = pod_infos.get("nodes")
-    if not isinstance(nodes, list) or len(nodes) == 0:
-        return None
-
-    node = nodes[0]
-    if not isinstance(node, dict):
-        return None
-
-    notebook_id = node.get("plotNotebookId")
-    workspace_id = node.get("accountId")
-    if notebook_id is None or workspace_id is None:
-        return None
-
-    local_storage = {
-        "plots.is_agent_controlled": "yes",
-        "viewAccountId": str(workspace_id),
-        "latch.authData": orjson.dumps({
-            "status": "done",
-            "auth0Data": {
-                "idToken": sdk_token.strip(),
-                "idTokenPayload": {
-                    "sub": "agent-session",
-                    "latch.bio/tos_ok": "true",
-                },
-            },
-        }).decode(),
-    }
-    return str(notebook_id), local_storage
 
 
 @trace_app_function_with_span
@@ -115,17 +57,6 @@ async def agent(s: Span, ctx: Context) -> HandlerResult:
         return "Agent not available"
 
     connection_idx += 1
-
-    try:
-        bootstrap = await _resolve_headless_browser_bootstrap()
-        if bootstrap is not None:
-            print("[agent endpoint] Starting headless browser")
-            notebook_id, local_storage = bootstrap
-            await start_headless_browser(notebook_id, local_storage=local_storage)
-        else:
-            raise RuntimeError("Failed to resolve headless browser bootstrap")
-    except Exception as e:  # noqa: BLE001
-        print(f"[agent endpoint] Failed to bootstrap headless browser: {e!s}")
 
     try:
         while True:
